pyproject.toml
[project]
name = "tailwind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-search-results>=2.4.2",
    "openai>=1.52.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "serpapi>=0.1.5",
    "streamlit-audiorecorder>=0.0.5",
    "streamlit-option-menu>=0.4.0",
    "streamlit>=1.39.0",
    "swarm",
]

[tool.uv.sources]
swarm = { git = "https://github.com/openai/swarm.git" }
